requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
python-dotenv==1.0.0
openai==1.0.0
PyYAML==6.0
requests==2.31.0
pydantic==1.10.7
pytest==7.4.0
